+++ cli/csstools-cli/src/cli.ts
@@ -26,6 +26,8 @@ import postcssOverflowShorthand from './plugins/postcss-overflow-shorthand';
 import postcssPlace from './plugins/postcss-place';
 import postcssPresetEnv from './plugins/postcss-preset-env';
 import postcssPseudoClassAnyLink from './plugins/postcss-pseudo-class-any-link';
+import postcssSteppedValueFunctions from './plugins/postcss-stepped-value-functions';
+import postcssUnsetValue from './plugins/postcss-unset-value';
 
 main();
 function main() {
@@ -114,6 +116,12 @@ function main() {
 		case 'postcss-pseudo-class-any-link':
 			postcssPseudoClassAnyLink();
 			return;
+		case 'postcss-stepped-value-functions':
+			postcssSteppedValueFunctions();
+			return;
+		case 'postcss-unset-value':
+			postcssUnsetValue();
+			return;
 		default:
 		{
 			const allHelp = [
@@ -153,6 +161,8 @@ function main() {
 				'  postcss-overflow-shorthand',
 				'  postcss-place',
 				'  postcss-pseudo-class-any-link',
+				'  postcss-stepped-value-functions',
+				'  postcss-unset-value',
 
 				'\nPlugin Help:',
 				'  @csstools/csstools-cli <plugin-name>',