+++ plugin-packs/postcss-preset-env/src/plugins/plugins-by-id.mjs
@@ -36,7 +36,9 @@ import postcssReplaceOverflowWrap from 'postcss-replace-overflow-wrap';
 import postcssPlace from 'postcss-place';
 import postcssPrefersColorScheme from 'css-prefers-color-scheme';
 import postcssColorRebeccapurple from 'postcss-color-rebeccapurple';
+import postcssSteppedValueFunctions from '@csstools/postcss-stepped-value-functions';
 import postcssFontFamilySystemUI from '../patch/postcss-system-ui-font-family.mjs';
+import postcssUnsetValue from '@csstools/postcss-unset-value';
 
 // postcss plugins ordered by id
 export const pluginsById = new Map(
@@ -79,6 +81,8 @@ export const pluginsById = new Map(
 		['place-properties', postcssPlace],
 		['prefers-color-scheme-query', postcssPrefersColorScheme],
 		['rebeccapurple-color', postcssColorRebeccapurple],
+		['stepped-value-functions', postcssSteppedValueFunctions],
 		['system-ui-font-family', postcssFontFamilySystemUI],
+		['unset-value', postcssUnsetValue],
 	],
 );